requirements.txt
# Core requirements without torch/transformers
# These are lighter and won't cause permission issues

# Core Streamlit
streamlit>=1.35.0

# LangChain ecosystem
langchain==0.2.11
langchain-core==0.2.41
langchain-community==0.2.10
langchain-text-splitters==0.2.4
langsmith==0.1.117

# LangChain integrations
langchain-google-genai==1.0.8
langchain-openai==0.1.23

# AI Model APIs
google-generativeai>=0.7.0
openai>=1.0.0
mistralai>=0.4.0
huggingface-hub>=0.23.0

# Vector Database
faiss-cpu==1.8.0

# Document Processing
pypdf>=4.0.0
python-docx>=1.0.0
openpyxl>=3.1.0
chardet>=5.0.0

# HTTP & Networking
requests>=2.31.0

# Environment & Utils
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0

# Note: Install torch separately if needed:
# pip install torch --index-url https://download.pytorch.org/whl/cpu
# Or for CUDA:
# pip install torch --index-url https://download.pytorch.org/whl/cu118